Add FastArrayAppendInst
Summary:
Add an instruction for appending one `FastArray` onto another. Later
diffs will speed up this operation specifically, but for now it is just
implemented on top of `push`.

Also note that the current implementation will modify the target array
even if pushing all of the elements results in an exception (since each
element is pushed one-by-one). In practice, this doesn't matter for now.

Reviewed By: avp

Differential Revision: D44235206

fbshipit-source-id: ae4b4cda0789f11b249022cab278d8e453c5222e