handle React.ChildrenArray in ts translation
Summary:
this isn't 100% correct, but good enough for most cases. in flow this type is recursive: https://github.com/facebook/flow/blob/main/lib/react.js#L278
we don't yet have a good way to add a type like this during translation. added TODO for reminder to update this when we can.

Reviewed By: pieterv

Differential Revision: D44845991

fbshipit-source-id: deb53aa61989acb41e4d5b0307d18a2106374691