Add a visitor for parsing serialised literals
Summary:
Add the ability to visit serialised literals with a visitor passed in
as a template parameter. This results in a simpler API, since the
parsing can remain completely separate from VM details.

More importantly, this can also perform better, since the logic for
each user can be written to deal directly with the serialised value.
For example, if some code wants to produce `SmallHermesValue`s, it is
currently forced to first produce a `HermesValue`, and then convert it.
With the new API, it can directly produce a `SmallHermesValue`.

Reviewed By: avp

Differential Revision: D44262703

fbshipit-source-id: e78d7e710afef689a7ad585c04a8a76256dc237c