Stop using global capture information for frame store elimination
Summary:
Currently, if a variable does not have any loads in capturing
functions, we eliminate stores across calls, since those calls can't
read the stored value. However, this breaks in the presence of
exceptions, since throwing will cause control to leave the current
basic block, and the captured variable may be observed if the exception
is caught.

For now, scrap the analysis that uses information about capturing loads
globally. We can still do some analysis in the entry basic blocks, since
the entry block cannot be in a try, and variables owned by the current
function cannot be observed by enclosing functions.

Reviewed By: avp

Differential Revision: D44945582

fbshipit-source-id: c2c7d05081cfc836db1ba5f8ffac71502b978972