Add FastArrayLoadInst
Summary:
Add a special instruction for accessing elements in a `FastArray`. For
now, it performs relatively expensive checks on the index to make sure
it is an unsigned integer within the range of the array, so the
function is placed out-of-line. We may want to revisit the cost of the
checks (and therefore the semantics) or the inlining later on.

Reviewed By: avp

Differential Revision: D44241717

fbshipit-source-id: d8f381f6661776386c8a651a7a8afde7078beb00