prevent duplicate react import during ts translation
Summary: this definitely needs a rework in general, but to get the naive version of this working for now this diff adds a `null` state to the `shouldAddReactImport` flag, representing that we have not encountered React types to transform. the `shouldAddReactImport` flag will be set to false if a valid `React` reference is detected and, once the flag is set to false, it will not be re-checked. this prevents adding a duplicate react import if a global react type is used later.

Reviewed By: pieterv

Differential Revision: D44937972

fbshipit-source-id: 325d27bac073365e4162cfb774158ce844ddb7d1