Add a basic linux smoke test in CircleCI (#974)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/974

Add a simple CI job to run Static Hermes tests on linux.

This needs to use a newer version of Debian that mainline Hermes in
order to have a sufficiently new version of CMake available. However,
this latest version also updates to ICU72, which contains a breaking
change in the type of space used when formatting date strings.

Other engines have chosen to scan the result and replace the narrow
no-break space with a regular space. For now, just make our tests more
permissive.

Reviewed By: tmikov

Differential Revision: D45171190

fbshipit-source-id: c69fba4adc75c9989836db7d46f8f602402b0253