Parse conditional type
Summary:
This diff implements the conditional type parsing support. Infer type
parsing is more involved and requires backtracking, which I will do in
a future diff.

I'm using
https://github.com/swc-project/swc/blob/aaaa9caafcc5bf6b269bf6533e8c0198825cb33f/crates/swc_ecma_parser/src/parser/typescript.rs#L913-L958
as a reference. The `disallow_conditional_types` flag is only useful
for infer types, which will be set and handled in the infer type
parsing diff.

Changelog: [Internal]

Reviewed By: avp

Differential Revision: D44313906

fbshipit-source-id: 71a40cc33c85a84165110e416f61746d81cdc7e7