Switch linux tests to Ubuntu Jammy (#979)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/979

Debian Bookworm is not yet stable. I had originally used it because we
needed a new CMake and our tests were not passing under Ubuntu Jammy
(which is LTS and has a compatible CMake).

After some digging, the issue in Ubuntu was just that the container did
not come with `tzdata` preinstalled. Edit our test to install `tzdata`,
and switch to Ubuntu, to avoid issues from Debian Bookworm continuing
to change over time.

Differential Revision: D45290420

fbshipit-source-id: 70e84277fa863fcd6994905ab1531cae7eae781d