Implement declare component parsing
Summary:
Implement `declare component` parsing. e.g.

```
declare component Foo(bar: Bar, 'foo': Foo);
```
The main difference from the component syntax is the params are types
so local values are not needed or allowed:
- Unlike declare function param names are required since the name is
  significant.
- `as` is banned since there is not use case for this syntax here since
  the value cannot be used.
- String literal params are not required to be followed by an `as`.
- For rest params the param name is optional.

In terms of implementation i copied `FunctionTypeParam` for the AST
format and used the `parseEnumDeclarationFlow` style of passing a
`declare` enum to `parseComponentDeclarationFlow` so this logic could
be reused. Unfortunately due to the param passing differences i had to
reimplement all the param parsing logic which mean some duplicated
logic.

Reviewed By: avp

Differential Revision: D45012272

fbshipit-source-id: 8454a41c9354f29980ce323232bbfc3112c8aa65