Expose optional value on MemberExpression
Summary: Hermes parser uses `OptionalMemberExpression` but to match estree we convert the AST node to `MemberExpression` with an `optional = true` property. However since we codegen the member expression generator from the `ESTree.def` definition it was missing the optional property. As with other similar cases here i created a manual generator that correctly has the optional type.

Reviewed By: noahlemen

Differential Revision: D45337798

fbshipit-source-id: 25372c760a3be7a2a7ca81cfcc82c6163659b840