Require type annotation in ComponentTypeParameter
Summary:
To better align with the existing `FunctionTypeParam` node we have
decided to make the type annotation in `ComponentTypeParameter` non
optional. What this concretely means is the following code will now
error:
```
declare component Foo(bar);
                      ~~~^ ':' expected in component type parameter
```
Flow would require a type annotation in this position anyway and the
rules for `FunctionTypeParam` would have parsed `bar` as a type
annotation so since names are always required in component params its
better to be consistent and always require both a name and type
annotation (note `name` is optional on the `ComponentTypeParameter` node
since its reused with `rest` arguments where it is acceptable to not
define a name).

Reviewed By: avp

Differential Revision: D45588227

fbshipit-source-id: 29f3fda0cdbdaaf59c81efc27b49150b8d59031d