[Hermes] Handle dead Phis in simplifyPhiInst
Summary:
It is possible to have a Phi in a dead block of the form:
```
%15 = PhiInst %15 : undefined|number, %BB4
```

This is a Phi with a single entry, which means that the current logic
will attempt to replace all uses of the Phi with itself, and then
delete it. The end result is that all existing users of the Phi
continue to hold the now freed Phi as an operand.

This can be fixed by simply deleting the special case for single entry
Phis, since the subsequent loop is more general and more correct. It
only replaces Phis where all entries refer to the same (non-self)
value, and performs no optimisation in the case where all entries are
self-edges.

@already-on-github

Test Plan: .

Reviewers: tmikov, avp, fbmal7, mattblagden, #hermes-buddies

Reviewed By: avp

Subscribers: #hermes

Differential Revision: https://phabricator.intern.facebook.com/D45760271

Tasks: T152865156

Tags: hermes