Avoid resetting process.exitCode when initializing HermesParserWASM (#989)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/989

The default emscripten code overrides the `process.exitCode` value to `0` on module init. This means if someone has already set an `exitCode` it will be reset. This change fixes this by switching emscripten to use the `MODULARIZE` flag which allows us to pass in function overrides, we then pass a modified implementation of `quit` that only adds the exitCode if its non zero.

Fixes: https://github.com/facebook/hermes/issues/978

Reviewed By: tmikov

Differential Revision: D45705102

fbshipit-source-id: 915ef98348be0b6f6536ea902a8b62a10490a66b