[Hermes] Use dummy throw when throwing TypeError
Summary:
We currently have two places that invoke `HermesBuiltin.throwTypeError`
to produce a `TypeError`, but because that call isn't a terminator,
introduce a dummy return after the call to satisfy IR invariants.
However, even though the return is unreachable, it can have effects on
the IR in indirect ways.

For instance, in the attached test case, inlining turns the unreachable
return inside the `try` into a direct jump, without first emitting a
`TryEndInst`. This results in ill-formed IR, since we appear to never
leave the `try`. In the attached test, this results in the exception
thrown getting caught and printed by the catch in the inlined
function.

To fix this, replace the dummy return with a dummy throw. This is more
appropriate, since it is unaffected by inlining, and more directly
models the effects of throwing a `TypeError`.

@already-on-github

Test Plan: added a test

Reviewers: tmikov, avp, fbmal7, mattblagden, #hermes-buddies, #hermes

Reviewed By: avp

Subscribers: #hermes

Differential Revision: https://phabricator.intern.facebook.com/D45782748

Tasks: T152861317

Tags: hermes