Create prettier-plugin-hermes-parser package
Summary: Create a simple package that exposes `hermes-parser` as a Prettier plugin. This will allow the easy use of `hermes-parser` without needing to merge the parser into upstream prettier. The limitation is we have no control of printing logic but this package is already constrained to what the Flow parser supports anyway so it shouldn't be an issue.

Differential Revision: D45965450

fbshipit-source-id: 767136b73dfcf1bfbc82cdfd2fdb397eda2ceb77