[Hermes] Move simplifySwitchInst to SimplifyCFG
Summary:
`simplifySwitchInst` modifies the CFG and can create unreachable basic
blocks, which need to be handled properly in SimplifyCFG.

The attached test was creating IR like this in the middle of
InstSimplify (abridged):
```
%BB3:
  %15 = PhiInst %19 : undefined|number, %BB4
%BB4:
  %19 = PhiInst %15 : undefined|number, %BB3
```
and running `simplifyPhiInst` on that would result in
```
%BB3:
  %15 = PhiInst %15 : undefined|number, %BB4
```
because for single-operand Phi we replace uses of the Phi with the
operand of the Phi, resulting in invalid IR.
This happens in unreachable blocks because the Phis can simply dominate
each other.

Moving `simplifySwitchInst` to SimplifyCFG fixes this by ensuring that
InstSimplify itself does not create unreachable blocks.

Next step is to only iterate the reachable blocks in InstSimplify.

@already-on-github

Test Plan: add test

Reviewers: tmikov, neildhar, #hermes-buddies, #hermes

Reviewed By: neildhar

Subscribers: luigiconiglio, #hermes

Differential Revision: https://phabricator.intern.facebook.com/D45738345

Tags: hermes