Create prettier-plugin-hermes-parser package (#999)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/999

Create a simple package that exposes `hermes-parser` as a Prettier plugin. This will allow the easy use of `hermes-parser` without needing to merge the parser into upstream prettier. The limitation is we have no control of printing logic but this package is already constrained to what the Flow parser supports anyway so it shouldn't be an issue.

Reviewed By: SamChou19815

Differential Revision: D45965450

fbshipit-source-id: caa881448b91d6724159785433273288df929754