Explicitly specify prototype for excludedItems in destructuring
Summary:
We currently do not specify the prototype of the `excludedItems`
object, which means that it defaults to `Object.prototype`. Since
`Object.prototype` can be modified by the user, and we do not only
consider "own" properties when checking whether a property exists in
`excludedItems`, this means that properties on `Object.prototype` can
be wrongly excluded.

It also means that you could set `Object.prototype.__proto__` to a
proxy and violate VM expectations about `excludedItems`.

To fix this, always explicitly set the prototype.

Reviewed By: avp

Differential Revision: D46077434

fbshipit-source-id: f5d4d0f05d1d047d31bc7970ccfd1f0f05f881eb