Support prettier V2 with prettier plugin
Summary:
Modify the hermes prettier plugin to support prettier v2 but with prettier v3 printing logic. This enables prettier across all versions to support all hermes parser syntax.

To enable this logic we embed a modified version of prettier v3 within this repro that exposes the v3 `astToDoc` logic and we use this within the estree printer of this plugin.

Differential Revision: D46173843

fbshipit-source-id: 1d2654fe5aaf929ae66754159453000421e661e0