Support prettier V2 with prettier plugin (#1012)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1012

Modify the hermes prettier plugin to support prettier v2 but with prettier v3 printing logic. This enables prettier across all versions to support all hermes parser syntax.

To enable this logic we embed a modified version of prettier v3 within this repro that exposes the v3 `astToDoc` logic and we use this within the estree printer of this plugin.

Reviewed By: SamChou19815

Differential Revision: D46173843

fbshipit-source-id: 14d67b57c7660528d6ee2a14b79341d3df0ee60f