Add Value Attributes
Summary:
In order to be able to arbitrarily associate Values with attributes,
we move the storage into `Value`.
Soon we'll be able to put attributes on instructions.

For now, because there are less than 8 attributes,
we can just put them in `Value` directly.
Eventually when there's enough of them we can move the actual storage
into Module, so we've forced passing `Module` in when retrieving them.

Reviewed By: tmikov

Differential Revision: D45128870

fbshipit-source-id: 62c21c23e2c72affac4d17a83c333fea637d82a2