Avoid doCall when calling known NativeJSFunctions
Summary:
We know when we're compiling a NativeJSFunction,
so there's no reason to do the extra check in `doCall`.
We can just call the function directly if we know the target,
and if we don't, we can still skip the `doCall` and read the pointer
directly.

To allow this, add a new attribute `isNativeJSFunction` to `Call`
which is recognized during ISel and which bypasses `doCall`.

Reviewed By: tmikov

Differential Revision: D44889355

fbshipit-source-id: 00fc99ead4175cd2d8c24d3f3d56c7b85a7480e4