Always emit ThrowTypeErrorInst in IRGen
Summary:
We currently have two places that invoke `HermesBuiltin.throwTypeError`
to produce a `TypeError`, but because that call isn't a terminator,
introduce a dummy return after the call to satisfy IR invariants.
However, even though the return is unreachable, it can have effects on
the IR in indirect ways.

For instance, in the attached test case, inlining turns the unreachable
return inside the `try` into a direct jump, without first emitting a
`TryEndInst`. This results in ill-formed IR, since we appear to never
leave the `try`. In the attached test, this results in the exception
thrown getting caught and printed by the catch in the inlined
function.

To fix this, always emit`ThrowTypeErrorInst`, which is a terminator.

Reviewed By: avp

Differential Revision: D45915742

fbshipit-source-id: c782f58c4cba7216293fac4b24aeb34bf3daca33