Add Type-TypeInfo indirection
Summary:
In order to be able to easily replace types while doing resolution,
as well as associate extra information such as AST nodes with the types,
we separate the notion of "TypeInfo" (the backing storage for the Type)
from the `Type` itself.

For example, this allows us to easily swap out UnionType once it's
canonicalized for a single type if all the arms were the same.

This diff is a mechanical change that just adds `TypeInfo` and `Type`
as distinct classes, but doesn't actually modify any logic in the
checker to take advantage of the indirection.

There are two tests which fail due to the handling of structural
types in this diff, which was buggy prior to this diff anyways.
The next diff fixes these tests and removes the XFAIL tags.

Reviewed By: tmikov

Differential Revision: D45319910

fbshipit-source-id: b44656dec2176b7aa7199e9cef465ed9b581ba68