Make Type hashing shallow
Summary:
The same type can be represented in multiple ways, e.g.
```
type A = A[];

type B = C[];
type C = B[];
```
are all the same type and would have to hash to the same thing.

We're going to update comparison to handle this, but being able to
ensure they hash the same would be difficult.

Perhaps if we need deep hashing in the future we can revisit this.

For now, we use a shallow hash only based on the TypeKind.

The hash caching mechanism is still there to reduce calls to
`llvh::hash_value` and to allow the hash function to easily get more
expensive and deeper in the future.

Reviewed By: tmikov

Differential Revision: D45165949

fbshipit-source-id: 9667a3775d3655e58b91f38927bd3cbaf079aac7