Hardcode CommonJS as disabled
Summary:
The builtin implementation of CommonJS is not supported in this version
of Hermes, so make it impossible perform code generation with it.

Some parser parser tests still requires some minimal level of support,
so the change is a bit more surgical. The current flag accessed as
`getUseCJSModules()` is hard-coded as disabled, ensuring that no code
generation can occur.

At the same time, add a new flag `getTransformCJSModules()`, which is
only enabled when both `-commonjs` and `-dump-transformed-ast` are
passed (thus code generation is off!). This flag is only checked in one
place on SemanticValidator, performing AST transformation of exporting a
default function declaration.

Delete all code-generation related tests, but keep the parser related
ones.

Reviewed By: neildhar

Differential Revision: D46305732

fbshipit-source-id: 2cfb906a580acfcb404dff1b2dfda8b4416324a9