Relax criteria for accesses from exception blocks in Mem2Reg (#992)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/992

The criteria for promoting stack locations in exception blocks to SSA
is currently extremely conservative. Any block that is dominated by a
`TryStartInst` or a `CatchInst` is marked as unsafe for promotion.
This means that in the following code:
```
function foo(sink){
  var x = 0;
  try {} catch {}
  sink(x);
}
```
`x` will not be converted to an SSA value, since the line `sink(x)` is
dominated by a `TryStartInst`.

In general, only values that are stored to inside a `try` should be
marked as unsafe. This is because an exception thrown in a try may be
caught within the same function, so we cannot determine if the store
will actually occur, and therefore cannot represent the value obtained
by subsequent loads in SSA form.

A load in a try is safe to promote to SSA, since it just reads the
value stored before the try. A load/store in a catch is really no
different from any other load/store after the catch.

Reviewed By: tmikov

Differential Revision: D45846693

fbshipit-source-id: 4cc3171dc6029032456b7767d426ab8c2260fac9