Analyse try nesting depth to improve SHJmpBuf usage
Summary:
Add an analysis to determine the nesting depth of a given basic block
in try statements. This is useful for optimisations, where we need to
quickly determine if the exception thrown by some throwing instruction
may be caught in the current function, and observe local variables.

To start, use this in the backend to reuse `SHJmpBuf`s. Instead of
allocating an `SHJmpBuf` for every try, we can now only allocate
enough for the maximum nesting depth of try's in the function, and use
the one corresponding to the current nesting depth. This also lets us
pass in the `SHJmpBuf *` directly to `_sh_end_try`, which makes it
slightly more efficient to pop.

Reviewed By: tmikov

Differential Revision: D45584273

fbshipit-source-id: a8dbee1299d75f23dc670db14b092991245bd990