Add asserts to limit large segments
Summary:
Add an assert in the object creation path to more quickly detect if
an object is created with a size larger than the max.

Add a static assert in Hades when adding a new segment to the freelist
to verify that any region space in a segment can always be represented
in the size of a cell.

Reviewed By: tmikov

Differential Revision: D46416780

fbshipit-source-id: 65e052870228ca9d53638a3982b87a8786756944