Always serialize object literal keys
Summary:
Some literals (for instance a literal with only one element) do not pass
the serialisation criteria in `estimateBestNumElemsToSerialize`. Copy
the pass for lowering `AllocObjectLiteral` over into an SH specific
pass, without the limit on how many elements to serialise.

This is desirable because:

1. Serialising the keys has a performance benefit, since the hidden
class gets cached and does not need to be recomputed each time.
2. The current heuristics are based loosely on HBC bytecode
instructions, which don't really apply in SH. The relative size of
serialising a few more `null` values is small relative to the size of
native code.

Reviewed By: avp

Differential Revision: D45014094

fbshipit-source-id: 45269f7c1dc956db55c7227f83275c882bf62ed3