Remove unused recursive functions
Summary:
If a recursive function was known to call itself,
it counted as a user of itself and wouldn't be deleted even if its
creation was deleted.
This would result in it trying to load itself from the frame
despite it never having corresponding stores anywhere in the program.

Fix this by checking only users outside the function itself when
deleting unused functions.

Reviewed By: tmikov

Differential Revision: D46622618

fbshipit-source-id: 716df1c41a25a65b5f29e554a09cd9aba6a064d0