Fix setting result type of HBCLoadFromEnvironmentInst
Summary:
We do not propagate the result type `LoadFrameInst` to
`HBCLoadFromEnvironmentInst` during lowering. This results in the type
being incorrect, since it is always `any`, regardless of whether the
actual type is narrower, or potentially empty.

It is likely this would break under more aggressive optimisations on
lowered IR.

To fix this, set the type of the instruction at construction to the
type of the variable, similar to what we do for `LoadFrameInst`. In
addition, when we translate from a `LoadFrameInst` to an
`HBCLoadFromEnvironmentInst`, set the type in the caller.

This should just be a temporary measure, since we are planning to add a
side-table, and since the lowered frame instructions are likely to be
deleted soon.

Reviewed By: avp

Differential Revision: D46725946

fbshipit-source-id: e206ed80eb4b13966421359a4e5860efb083d44c