Make register stack size configurable
Summary:
fewer-gc-bug0.js requires a large stack to run, because it invokes
`bind` with a large number of arguments to test the maximum size of an
`ArrayStorage`. It is failing in SH because it uses the default register
stack size of 1MB, instead of bumping it to 8MB like the `hermes`
tool.

Make the register stack size configurable on the command line, and
bump it up for this test.

Also fix populating the default max heap size, which was incorrectly
being populated with the default initial heap size.

Reviewed By: tmikov

Differential Revision: D44962630

fbshipit-source-id: 3ec85e451231b5d192328335fb765af935b705f8