Refine new.target analysis in FunctionAnalysis
Summary:
We currently do set `allCallsitesKnown` to false as long as a function
accesses `new.target` at all. However, this is pessimistic for
functions that are never invoked as constructors, since their
`new.target ` property is never actually the closure.

Refine the analysis by flipping it around. Instead of checking whether
the function uses `new.target`, look at each callsite to see if it
actually passes in the closure as the `new.target` operand. If it does,
check if the function actually uses that operand, and only if both of
those are true, mark the closure as having escaped.

This can have surprising effects, since arrow functions capture
the `new.target` of their enclosing scope, many functions are
initially generated as accessing `new.target`. Ideally, these
accesses are optimised away by other optimisations, but as
seen in the bigint-type-inference test modified by this diff,
not all cases are caught.

Reviewed By: tmikov

Differential Revision: D45205811

fbshipit-source-id: 9296401467d1ff13ad40576f86a921a11920891c