ComponentDeclaration don't output empty props
Summary: When building the `ComponentDeclaration` Babel adapter i originally thought downstream babel transforms would need the props object to exist to detect if a function is a component or not. However after further investigation i believe this not to be the case components without props is common. So this change switches the adapter to not output any function parameters if non are defined, this removes the overhead of having an unnecessary empty destructured object at runtime.

Reviewed By: jbrown215

Differential Revision: D46908383

fbshipit-source-id: 8ad7e37c743068d0d6a430fe7d195081e9a20225