[Hermes] Create 'this' param explicitly instead of as a named param
Summary:
The `Parameter` constructor was looking for the string `"this"` to
determine when to `setThisParameter()` on the parent function.
In reality, the caller knows when they are trying to create a `"this"`
parameter, so make a new `IRBuilder::createThisParameter` and be
explicit when creating 'this' parameters.

@already-on-github

Test Plan: add a test

Reviewers: tmikov, neildhar, #hermes-buddies, #hermes

Reviewed By: neildhar

Subscribers: #hermes

Differential Revision: https://phabricator.intern.facebook.com/D46913957

Tasks: T154991249, T149387898

Tags: hermes