Update Intl API documentation to include iOS support (#1042)
Summary:
As discussed in issue https://github.com/facebook/hermes/issues/1035,
the documentation of the Intl APIs does not include the current state
of the iOS support. This PR adds this information and formats the page
to reflect that this page now talks about Android and iOS support and
not just Android. Also the "Not yet supported" sections has been
removed. The page now only mentions what **is** supported and the
limitations of the functionalities that are currently supported. This
makes more sense with regard to the still changing specification.
Mentioning what is supported automatically means the remaining
information from the spec that is not included under "supported" is not
currently supported.

Pull Request resolved: https://github.com/facebook/hermes/pull/1042

Reviewed By: avp

Differential Revision: D47155521

Pulled By: neildhar

fbshipit-source-id: c30c51008b734d922aeb5856e805642a3657a8c0