Add LocalNonPtr RegClass and multi-class RegisterFile
Summary:
Adds the ability for the `RegisterFile` to track all register classes.
Instead of just using one `BitVector`, we use one per class,
and ask users of the `RegisterFile` to provide the class for each call.

The other option is to have multiple `RegisterFile`s, but it seemed
easier to put all this in a single `RegisterFile`.

Technically this adds an extra empty `BitVector` for `RegStack` which is
never used, but it also allows us to easily add more classes in the
future without having to have a manual `RegClass` to array index
conversion.

No functionality changes.
This just allows us to allocate `LocalNonPtr` in the future,
but doesn't do anything with it yet.

Reviewed By: tmikov

Differential Revision: D47134516

fbshipit-source-id: d51a998432a11d5832fff7a30578c80030085431