Fix SwitchLowering
Summary:
When handling the last case in a switch, both branches may lead to the
default block. e.g. in

```
function foo(i) {
  switch (i) {
    case 0:
      i = 2;
    case 1:
      break;
  }
  return i;
}
```

When we lower the last case to a conditional branch, both sides of that
branch lead to the same block. Because of this, `SwitchLowering`
currently incorrectly creates two entries for `BB4` in the `PhiInst` at
the start of `BB2`:

```
%BB0:
  %0 = HBCLoadConstInst (:number) 2: number
  %1 = LoadParamInst (:any) %i: any
  %2 = BranchInst %BB1
%BB2:
  %3 = PhiInst (:any) %0: number, %BB3, %1: any, %BB4, %1: any, %BB4
  %4 = ReturnInst %3: any
%BB3:
  %5 = BranchInst %BB2
%BB4:
  %6 = HBCLoadConstInst (:number) 1: number
  %7 = CmpBrStrictlyEqualInst %6: number, %1: any, %BB2, %BB2
%BB1:
  %8 = HBCLoadConstInst (:number) 0: number
  %9 = CmpBrStrictlyEqualInst %8: number, %1: any, %BB3, %BB4
```

To fix this, avoid copying Phi entries if the default block and the
destination block for the final case are the same.

Reviewed By: avp

Differential Revision: D46788088

fbshipit-source-id: 3e26364abfbd39ee0d6f9484d27e7e6314d68bad