Enforce that Phis have no duplicate entries
Summary:
The current IRVerifier check enforces that Phis never have two entries
for the same block that have different values. This invariant can be
tricky to correctly maintain.

For example, if we start out with a `Phi` that has two entries for a
block which both point to `LiteralUndefined`, `LoadConstants` will
create two separate `HBCLoadConstInst`s for the two, and the invariant
will get silently violated.

To avoid this, it is simpler to tighten the invariant such that each
block must only appear once.

Reviewed By: avp

Differential Revision: D46906390

fbshipit-source-id: fa01c01fb80c1787bb9296b7fd8899ec05273a9f