Fix exponent overflow in BigInt exponentiation
Summary:
The BigInt exponentiation code currently truncates the first digit of a
BigInt from 64 to 32 bits. This means that for a single digit BigInt
exponent that has a value larger than `UINT32_MAX`, we end up
truncating the exponent and producing an incorrect value. To fix this,
do not truncate the exponent until after we have guaranteed that it
will fit in 32 bits.

Reviewed By: fbmal7

Differential Revision: D47275000

fbshipit-source-id: 881b7653d0394b88aad9db11fb6f99e675a7e38f