Pass string to JSProxy::defineProperty when appropriate
Summary:
In one codepath in `putNamedWithReceiver_RJS`, we unconditionally treat
the key as a `Symbol` value when passing it to the methods
`JSProxy::defineProperty` and
`JSObject::getOwnComputedPrimitiveDescriptor`. However, other callers
of these functions first perform a check to verify if it's a Symbol or
string primitive. So, add this same check in `putNamedWithReceiver_RJS`.

Reviewed By: neildhar

Differential Revision: D47484057

fbshipit-source-id: 15efd9cd9f9f2c3ecab9e12f2dcb9e2d04f834f9