Fix transform print caching
Summary: This cache key is used to work around logic in prettier that causes it to reuse an existing parse function if the shape of a new plugin is the same. The cache key currently uses an incrementing count but prettier's caching seems to match across workers making this approach insufficient. To make this more robust we can create a unique base key for each instantiation of this code.

Reviewed By: kassens

Differential Revision: D47677844

fbshipit-source-id: b16d2f0d91d9500f0a05d102003607ac2fbcc62c