prevent duplicate react import during ts translation
Summary:
this definitely needs a rework in general, but to get the naive version of this working for now this diff adds a `null` state to the `shouldAddReactImport` flag, representing that we have not encountered React types to transform. the `shouldAddReactImport` flag will be set to false if a valid `React` reference is detected and, once the flag is set to false, it will not be re-checked. this prevents adding a duplicate react import if a global react type is used later.

Original Author: noahlemen@meta.com
Original Git: 73f828f0b2d218ab5f2a6d8a7b66ae6fbddc36c6

Original Reviewed By: pieterv

Original Revision: D44937972

Reviewed By: neildhar

Differential Revision: D47690093

fbshipit-source-id: b08c61dc9b431cd27e96b46a1101ceddb0bc547e