Strengthen invariant for SNOP list in LowerAllocObject
Summary:
We should stop processing `StoreNewOwnPropertyInst`s as soon as we find
a usage of the `allocObject_` where it is not the target of a SNOP. To
keep the invariant simple this should include the case where an object
is stored as a property on itself.

I don't think it's actually possible to emit this IR currently, but
this clarifies the invariant.

This invariant is also convenient for the next diff, since it means
that each serialisable SNOP may only account for a single use of an
`AllocObjectInst`.

Reviewed By: avp

Differential Revision: D30649306

fbshipit-source-id: 5f85edb7735cba34ca2ec73af2166596825632ac