Switch hermes-transform print to use the hermes parser prettier plugin
Summary:
Prettier does not have an API to provide an AST to format. So to work around this limitation we use the `parser` API but prettier hardcodes this to use the `ASTFormat` value of `ESTree` which means it wont print via the `prettier-plugin-hermes-parser` plugin even if its installed and configured. Unfortunately there is no way to dynamically define the printer since prettier doesn't actually know what one to use as it resolve the printers via the filename which is not provided here. So since all the cases we know of are already using the `prettier-plugin-hermes-parser` this change just hard codes it as the printer. The one exception to this is the `flow-api-translator` when printing TS, however the printing differences are minimal and it should only be used for generated code so this seems like a reasonable tradeoff to upgrade this case to the latest printing. If it really becomes an issue we can allow a hard coded "parser: 'hermes' | 'typescript'" value to be passed to support both.

Original Author: pieterv@meta.com
Original Git: e9c731912ee99d9fc4133ed9c4a0b7a87744dc06

Original Reviewed By: noahlemen, SamChou19815

Original Revision: D46922228

Reviewed By: neildhar

Differential Revision: D47690628

fbshipit-source-id: 39b7091d9eb40215915ddde010a12af503523003