Run HermesRuntimeTests with all runtime types
Summary:
Use the same runtime factory we use for JSI tests when running the
`HermesRuntimeTest`, so the tests are run on any runtime that is built
on top of Hermes. For now, this exercises `HermesRuntime` and the
"thread-safe" `HermesRuntime`. This will be expanded to include the
upcoming ABI-safe runtime and WASM based runtime.f

Reviewed By: avp

Differential Revision: D47379092

fbshipit-source-id: 83bd1fcfd2dac7cd32ade6ec2b76c397501cb7a7