Move log buffer warning to the start
Summary:
I had the warning showing at the end and was thinking maybe it'd be better at the start instead.
{F1063051011}

If developers don't need the earlier messages, then they don't have to see the warning.
If they scroll to the top looking for the earlier messages but don't see them, then having that warning at the start would tell them why.

Reviewed By: mattbfb

Differential Revision: D47998805

fbshipit-source-id: 8f3d353243470697c1493f22b49c527c3b4fb301