Bump Android Gradle Plugin to 7.1 (#1085)
Summary:
The version of AGP used inside Hermes is quite old (4.1.0)
I've bumped it ahead to a more recent one (7.1) and removed the AndroidManifest.xml file which is causing a build warning on React Native's end.

Pull Request resolved: https://github.com/facebook/hermes/pull/1085

Test Plan:
CI should be sufficient to test this one.
Specifically I had to tweak the emulator setup as the APK was failing to install with insufficient storage

Reviewed By: mdvacca

Differential Revision: D48149886

Pulled By: cortinico

fbshipit-source-id: 4df8450ebe278fd647a333bf50d6b711c55fb611