Update publish-tag React Native version description (#1086)
Summary:
When working on React Native releases and needing to do a Hermes bump, we need to run the the publish-tag workflow. It requires the React Native version that will use the Hermes tag that this workflow will publish.

The current phrasing doesn't make this clear. This is my attempt to make this version description clearer.

Pull Request resolved: https://github.com/facebook/hermes/pull/1086

Test Plan: N/A

Reviewed By: fbmal7

Differential Revision: D48167473

Pulled By: mattbfb

fbshipit-source-id: 4ce2a2ef76d788d55f4ae377bf0e36a6e8ba3a54