Impose a limit on native recursion in regex match
Summary:
`Context<Traits>::match` is a recursive function. However, we don't
have any checks for the depth that we are recursing. To fix this, we
simply add a member to the `Context` variable that is passed between
recursive calls. This keeps track of how many native recursions are
allowed. We set this to 64 initially.

Also added in a missing error check to the recursive call of `match`.

Reviewed By: neildhar

Differential Revision: D46299990

fbshipit-source-id: 8916a27a64786ea058836a05598b11d9ed08bf1a