Revert D46299990: Multisect successfully blamed "D46299990: [hermes] Impose a limit on native recursion in regex match" for test or build failures
Summary:
This diff is reverting D46299990
D46299990: [hermes] Impose a limit on native recursion in regex match by fbmal7 has been identified to be causing the following test or build failures:

Tests affected:
- [hermes/service/tests:remote_runner_dev_test - main](https://www.internalfb.com/intern/test/562950034203848/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2711910
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: fbmal7

Differential Revision: D48182602

fbshipit-source-id: e19764e9e487d7bf2da8cb57afc8867a64e14e65