Link inspector lib in libhermes
Summary:
React Native links with the `libhermes` target. There are a few options
for exposing the CDP handler externally:
1. Bundle it with libhermes
1. Create another library purely for the CDP handler

Pros of going with option 1 is ease of integration for consumers. Cons
of option 1 is that if someone doesn't want the CDP handler, then
bundling together could increase app size.

The pros of option 2 is that it's a separate library and consumers can
optionally include it and manage app size that way.

Since there is already precedence of using `HERMES_ENABLE_DEBUGGER` to
control the inclusion of `DebuggerAPI.cpp` functionality, we could use the
same mechanism to ensure CDP handler isn't included when debugger
functionality isn't there. Thus mitigating the app size concern. This
seems to be the best of both worlds: ease of use + not increasing app
size for release builds

Reviewed By: tmikov

Differential Revision: D48324876

fbshipit-source-id: 7e73e3a87cd4abee22981d74807825b8ee74a148