Infer types for the logical expressions
Summary:
`&&`, `||`, and `??` need their types inferred.

Importantly, `||` and `??` should exclude null/void if they only appear
on the LHS, because they necessarily can't be the result of the
expression (the types are always `false`).
To allow this, add an `excludes` set to `maybeCreateUnion` to let the
user control the final types prior to canonicalization step.

Reviewed By: tmikov

Differential Revision: D47025269

fbshipit-source-id: 2fc60b985e3414937419bb1fb337280c16e1efca