Convert HermesParserBabelAdapter to work on ESTree AST
Summary:
Previously the `hermes-parser` adapters both worked on the Hermes AST like this:
- Hermes AST -> ESTree AST
- Hermes AST -> Babel AST

This change switches it so we always convert to an ESTree AST then if `babel: true` is set we convert this to a Babel AST, e.g.
- Hermes AST -> ESTree AST
- Hermes AST -> ESTree AST -> Babel AST

This makes the babel lowering step safer as we can rely on the estree types to provide type safety (see issues fixed in this change) but also allows us to split out lowering steps such as Flow and ReactScript from the babel AST transform. This makes these transforms safer while making the code easier to understand since they are separate steps. e.g. now we have:

- Hermes AST -> ESTree AST
- Hermes AST -> ESTree AST -> ReactScript -> Flow strip -> Babel AST

Reviewed By: SamChou19815

Differential Revision: D48174633

fbshipit-source-id: d129373db2cad9e4e8c39856e5dc4256a5f67660