Split LowerAllocObjectFuncContext
Summary:
`LowerAllocObjectFuncContext` currently loosely operates in three
steps:

1. Construct the map for a given `allocInst` containing the users in
each basic block.
2. Produce an ordered list for each of the blocks that contains a user.
3. Traverse the ordered blocks to create an ordered list of
`StoreNewOwnPropertyInst`s that store to the `allocInst`.

Split these steps such that:
1. Is done by `runOnFunction`, and passed to `collectInstructions`.
2. Is factored out into a new helper `orderBlocksByDominance`, invoked
by `collectInstructions`. The helper is written to be reusable, so we
can factor it out into a utility file if it turns out to be useful in
other places.
3. Is done in `collectInstructions`.

While doing this split, it is also simple to move constructing the
`DominanceInfo` so that it is only done once per function, instead of
once per `AllocObjectInst`.

Reviewed By: tmikov

Differential Revision: D47724424

fbshipit-source-id: c45a468cf3799955f2589dae87c59dd2481ce44a