Improve scalability of LowerAllocObject
Summary:
`LowerAllocObject` can encounter poor performance when the program
contains basic blocks with a large number of object literals. This is
because for each object allocation, it does a pass over at least every
instruction in the basic blocks that contain users we want to evaluate
for lowering. For a large nested JSON literal, thousands of objects may
be in the same basic block, so iterating over this huge basic block for
each object allocation quickly becomes expensive.

However, it is not necessary to do a full pass for every object
allocation. Instead, we can construct an ordered list of the stores to a
given allocation in each basic block with a single pass over the
function. Then, for each `AllocObject`, we can traverse just its stores.

Reviewed By: tmikov

Differential Revision: D30626992

fbshipit-source-id: 4c68ec8eb9cec40a29bc2c46662c688b3860faaa