Emit LoadConstant as needed
Summary:
We currently emit all `LoadConstant`s at the top of a function. In
large functions, particularly large object literals, this can create a
huge amount of register pressure in debug mode, since the literals are
not moved closer to the point where they are used.

To avoid this, always emit `LoadConstant` at the point where it is
needed. When optimisations are enabled, we can then rely on
CodeMotion to hoist and CSE to deduplicate them.

Reviewed By: tmikov

Differential Revision: D43048358

fbshipit-source-id: 7add8d6393f3ade22342e9e37d142d578cce3d1a