Allow VisitCaller to send parents to the Node overload of visit
Summary:
If the user only had the
```
void visit(Node *node, Node *parent);
```
overload in the visitor, it couldn't be called from the RecursiveVisitor
because the RecursiveVisitor was only looking for overloads for each
specific `NodeKind`.

Switch from using `static_cast` to `std::declval` and attempting to just
call the `visit` function directly inside `decltype`.

This allows the full overload resolution to take place.

Reviewed By: tmikov

Differential Revision: D48665671

fbshipit-source-id: 72c1a974412c2b3527dac4e135c8ba988745a5fb