Allow VisitCaller to send parents to the Node overload of visit
Summary:
If the user only had the
```
void visit(Node *node, Node *parent);
```
overload in the visitor, it couldn't be called from the RecursiveVisitor
because the RecursiveVisitor was only looking for overloads for each
specific `NodeKind`.

Switch from using `static_cast` to `if constexpr` on a
`has_parent_param` function.

This allows the full overload resolution to take place.

Reviewed By: tmikov

Differential Revision: D48665671

fbshipit-source-id: c24e69a154c679b7240768383e87172dbe39ea46