Track and typecheck implicit returns
Summary:
After running SemanticResolver for a function,
run a second pass to find statements that can terminate the function and
track them so that we know when every path through the function has one.
Determines whether or not the implicit return `undefined` is
reachable for that function.
Store that as a flag in the FunctionInfo and check it at typecheck
time.

Reviewed By: tmikov

Differential Revision: D47071429

fbshipit-source-id: 034c5042b9e20522914578073424896e7aafbb20