Handle serialization of literals with __proto__
Summary:
Instead of disqualifying literals with a parent specified, allow them
to be serialized by setting the prototype in an explicit second step
(when it is profitable to do so).

This will make it easier to use the new object literal mechanism for
the excludedItems in destructuring, since they need to have their
prototypes set to null.

Reviewed By: tmikov

Differential Revision: D48658154

fbshipit-source-id: a7c9a6cd2d2ec7eaadafb41b4e415382617722fb